+++ x/uibc/uics20/memo_handler_test.go
@@ -19,7 +19,6 @@ import (
 func TestValidateMemoMsg(t *testing.T) {
 	assert := assert.New(t)
 	receiver := accs.Alice
-	wrongSignerErr := "signer doesn't match the receiver"
 	asset := coin.New("atom", 10)
 	assetH := coin.New("atom", 5)
 	asset11 := coin.New("atom", 11)
@@ -27,7 +26,7 @@ func TestValidateMemoMsg(t *testing.T) {
 	goodMsgSupply := ltypes.NewMsgSupply(receiver, asset)
 	goodMsgSupply11 := ltypes.NewMsgSupply(receiver, asset11)
 	goodMsgSupplyColl := ltypes.NewMsgSupplyCollateral(receiver, asset)
-	goodMsgSupplyCollH := ltypes.NewMsgSupplyCollateral(receiver, assetH)
+	// goodMsgSupplyCollH := ltypes.NewMsgSupplyCollateral(receiver, assetH)
 	goodMsgSupplyColl11 := ltypes.NewMsgSupplyCollateral(receiver, asset11)
 	goodMsgBorrow := ltypes.NewMsgBorrow(receiver, asset)
 	goodMsgBorrowH := ltypes.NewMsgBorrow(receiver, assetH)
@@ -35,37 +34,48 @@ func TestValidateMemoMsg(t *testing.T) {
 	goodMsgLiquidate11 := ltypes.NewMsgLiquidate(receiver, accs.Bob, asset11, "uumee")
 	msgSend := &banktypes.MsgSend{FromAddress: receiver.String()}
 
+	errManyMsgs := "memo with more than 1 message is not supported"
+	errNoSubCoins := errNoSubCoins.Error()
+	errMsg0type := errMsg0Type.Error()
+	// errWrongSigner := "signer doesn't match the receiver"
+
 	mh := MemoHandler{leverage: mocks.NewLvgNoopMsgSrv()}
 	tcs := []struct {
 		msgs   []sdk.Msg
 		errstr string
 	}{
-		{[]sdk.Msg{ltypes.NewMsgSupply(accs.Bob, asset)}, wrongSignerErr},
-		{[]sdk.Msg{ltypes.NewMsgSupplyCollateral(accs.Bob, asset)}, wrongSignerErr},
+		/** we don't check signers in handlers v1
+		{[]sdk.Msg{ltypes.NewMsgSupply(accs.Bob, asset)}, errWrongSigner},
+		{[]sdk.Msg{ltypes.NewMsgSupplyCollateral(accs.Bob, asset)}, errWrongSigner},
 		{[]sdk.Msg{goodMsgSupplyColl,
-			ltypes.NewMsgBorrow(accs.Bob, asset)}, wrongSignerErr},
+			ltypes.NewMsgBorrow(accs.Bob, asset)}, errWrongSigner},
+		*/
 
 		// good messages[0]
 		{[]sdk.Msg{goodMsgSupply}, ""},
 		{[]sdk.Msg{goodMsgSupplyColl}, ""},
-		{[]sdk.Msg{goodMsgLiquidate}, ""},
+		{[]sdk.Msg{goodMsgLiquidate}, ""}, // in handlers v2 this will be a good message
 
 		// messages[0] use more assets than the transfer
-		{[]sdk.Msg{goodMsgSupply11}, "message must use only coins sent from the transfer"},
-		{[]sdk.Msg{goodMsgSupplyColl11}, "message must use only coins sent from the transfer"},
-		{[]sdk.Msg{goodMsgLiquidate11}, "message must use only coins sent from the transfer"},
+		{[]sdk.Msg{goodMsgSupply11}, errNoSubCoins},
+		{[]sdk.Msg{goodMsgSupplyColl11}, errNoSubCoins},
+		{[]sdk.Msg{goodMsgSupplyColl11}, errNoSubCoins},
+		{[]sdk.Msg{goodMsgLiquidate11}, errNoSubCoins},
 
 		// wrong message types
-		{[]sdk.Msg{goodMsgBorrow}, msg0typeErr},
-		{[]sdk.Msg{msgSend}, msg0typeErr}, // bank msg
-		{[]sdk.Msg{ltypes.NewMsgDecollateralize(receiver, asset)}, msg0typeErr},
-		{[]sdk.Msg{goodMsgBorrow, goodMsgBorrow}, msg0typeErr},
-		{[]sdk.Msg{goodMsgBorrow, goodMsgSupplyColl}, msg0typeErr},
-		{[]sdk.Msg{&ltypes.MsgLeveragedLiquidate{Liquidator: receiver.String()}}, msg0typeErr},
-		{[]sdk.Msg{goodMsgSupplyColl, goodMsgSupplyColl}, msg1typeErr},
-		{[]sdk.Msg{goodMsgSupplyColl, goodMsgLiquidate}, msg1typeErr},
-		{[]sdk.Msg{goodMsgSupplyColl, msgSend}, msg1typeErr},
+		{[]sdk.Msg{goodMsgBorrow}, errMsg0type},
+		{[]sdk.Msg{msgSend}, errMsg0type}, // bank msg
+		{[]sdk.Msg{ltypes.NewMsgDecollateralize(receiver, asset)}, errMsg0type},
+		{[]sdk.Msg{&ltypes.MsgLeveragedLiquidate{Liquidator: receiver.String()}}, errMsg0type},
 
+		{[]sdk.Msg{goodMsgBorrow, goodMsgBorrow}, errManyMsgs},
+		{[]sdk.Msg{goodMsgBorrow, goodMsgSupplyColl}, errManyMsgs},
+		{[]sdk.Msg{goodMsgSupplyColl, goodMsgSupplyColl}, errManyMsgs},
+		{[]sdk.Msg{goodMsgSupplyColl, goodMsgLiquidate}, errManyMsgs},
+		{[]sdk.Msg{goodMsgSupplyColl, msgSend}, errManyMsgs},
+		{[]sdk.Msg{goodMsgSupplyColl, goodMsgBorrow}, errManyMsgs},
+
+		/** uncomment when msg borrow enabled
 		// check msg borrow is after supply collateral
 		{[]sdk.Msg{goodMsgSupply, goodMsgBorrow}, "MsgBorrow must use MsgSupplyCollateral"},
 		{[]sdk.Msg{goodMsgSupply, goodMsgBorrowH}, "MsgBorrow must use MsgSupplyCollateral"},
@@ -74,14 +84,12 @@ func TestValidateMemoMsg(t *testing.T) {
 		{[]sdk.Msg{msgSend, goodMsgBorrow}, msg0typeErr},
 		{[]sdk.Msg{goodMsgSupplyCollH, goodMsgBorrowH}, ""},
 		{[]sdk.Msg{goodMsgSupplyColl, goodMsgBorrow}, ""},
+		*/
 
 		// more than 2 messages
-		{[]sdk.Msg{goodMsgSupplyColl, goodMsgBorrowH, goodMsgBorrowH},
-			"memo with more than 2 messages are not supported"},
-		{[]sdk.Msg{goodMsgSupplyColl, goodMsgLiquidate, goodMsgBorrow},
-			"memo with more than 2 messages are not supported"},
-		{[]sdk.Msg{goodMsgSupplyColl, goodMsgBorrowH, goodMsgBorrowH, goodMsgBorrowH},
-			"memo with more than 2 messages are not supported"},
+		{[]sdk.Msg{goodMsgSupplyColl, goodMsgBorrowH, goodMsgBorrowH}, errManyMsgs},
+		{[]sdk.Msg{goodMsgSupplyColl, goodMsgLiquidate, goodMsgBorrow}, errManyMsgs},
+		{[]sdk.Msg{goodMsgSupplyColl, goodMsgBorrowH, goodMsgBorrowH, goodMsgBorrowH}, errManyMsgs},
 	}
 
 	for i, tc := range tcs {